добавлена сборка контрактов

+++ Source/MiniJournal.Application/ArticlesService.cs
@@ -1,6 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
+using Infotecs.MiniJournal.Contracts.ArticlesApplicationService;
+using Infotecs.MiniJournal.Contracts.ImagesApplicationsService;
 using Infotecs.MiniJournal.Domain.Articles;
 using Infotecs.MiniJournal.Domain.Comments;
 using Infotecs.MiniJournal.Domain.Users;
@@ -23,6 +27,7 @@ namespace Infotecs.MiniJournal.Application
         private readonly IUserDomainService userService;
         private readonly IImagesService imagesService;
         private readonly ILogger logger;
+        private readonly IMapper mapper;
 
         public ArticlesService(
             IUnitOfWork unitOfWork,
@@ -33,7 +38,8 @@ namespace Infotecs.MiniJournal.Application
             ICommentDomainService commentService,
             IUserDomainService userService,
             IImagesService imagesService,
-            ILogger logger)
+            ILogger logger,
+            IMapper mapper)
         {
             this.unitOfWork = unitOfWork;
             this.articleFactory = articleFactory;
@@ -44,84 +50,108 @@ namespace Infotecs.MiniJournal.Application
             this.userService = userService;
             this.imagesService = imagesService;
             this.logger = logger;
+            this.mapper = mapper;
         }
 
         /// <summary>
         /// Возвращает список всех имеющихся статей
         /// </summary>
         /// <returns>Список всех статей</returns>
-        public Task<IEnumerable<Article>> GetArticlesAsync()
+        public async Task<GetArticlesResponse> GetArticlesAsync(GetArticlesRequest request)
         {
-            return this.articleRepository.GetArticlesAsync();
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            IEnumerable<Article> articles = await this.articleRepository.GetArticlesAsync();
+
+            return new GetArticlesResponse
+            {
+                Articles = this.mapper.Map<List<Contracts.ArticlesApplicationService.Entites.Article>>(articles)
+            };
         }
 
         /// <summary>
         /// Создать статью с указанным содержимым.
         /// </summary>
-        /// <param name="text">Содержимое статьи.</param>
-        /// <param name="image">Картинка.</param>
-        /// <param name="userId">Идентификатор пользователя создавшего статью.</param>
-        public async Task CreateArticleAsync(string text, byte[] image, long userId)
+        /// <param name="request">Запрос создания статьи.</param>
+        public async Task<CreateArticleResponse> CreateArticleAsync(CreateArticleRequest request)
         {
-            var user = await this.userService.GetUserByIdAsync(userId);
-            var imageId = await this.imagesService.UploadImageAsync(image);
-            var article = await this.articleFactory.CreateAsync(text, imageId, user);
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            var user = await this.userService.GetUserByIdAsync(request.UserId);
+            var createImageResponse = await this.imagesService.UploadImageAsync(new UploadImageRequest { Image = request.Image });
+            var article = await this.articleFactory.CreateAsync(request.Text, createImageResponse.ImageId, user);
 
             await this.articleService.CreateArticleAsync(article);
 
             await this.unitOfWork.SaveChangesAsync();
 
-            this.logger.Verbose("user {UserId} created new article: {ArticleText}", userId, article.Text.Truncate(30, true));
+            this.logger.Verbose("user {UserId} created new article: {ArticleText}", request.UserId, article.Text.Truncate(30, true));
+
+            return  new CreateArticleResponse(true);
         }
 
         /// <summary>
         /// Удаляет статью.
         /// </summary>
-        /// <param name="articleId">Идентификатор статьи.</param>
-        public async Task DeleteArticleAsync(long articleId)
+        /// <param name="request">Запрос удаления статьи.</param>
+        public async Task<DeleteArticleResponse> DeleteArticleAsync(DeleteArticleRequest request)
         {
-            var article = await this.articleService.GetArticleByIdAsync(articleId);
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            var article = await this.articleService.GetArticleByIdAsync(request.ArticleId);
             await this.articleService.DeleteArticleAsync(article);
 
             await this.unitOfWork.SaveChangesAsync();
 
-            this.logger.Verbose("article {ArticleId} was deleted", articleId);
+            this.logger.Verbose("article {ArticleId} was deleted", request.ArticleId);
+
+            return new DeleteArticleResponse(true);
         }
 
         /// <summary>
         /// Добавляет комментарий к статье.
         /// </summary>
-        /// <param name="text">Содержимое комментария.</param>
-        /// <param name="userId">Идентификатор пользователя добавившего комментарий.</param>
-        /// <param name="articleId">Идентификатор статьи, к которой добавляется комментарий.</param>
-        public async Task AddCommentAsync(string text, long userId, long articleId)
+        /// <param name="request">Запрос добавления статьи.</param>        
+        public async Task<AddCommentResponse> AddCommentAsync(AddCommentRequest request)
         {
-            var user = await this.userService.GetUserByIdAsync(userId);
-            var article = await this.articleService.GetArticleByIdAsync(articleId);
-            var comment = await this.commentFactory.CreateAsync(text, user, article);
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            var user = await this.userService.GetUserByIdAsync(request.UserId);
+            var article = await this.articleService.GetArticleByIdAsync(request.ArticleId);
+            var comment = await this.commentFactory.CreateAsync(request.Text, user, article);
 
             await this.commentService.AddCommentAsync(article, comment);
 
             await this.unitOfWork.SaveChangesAsync();
 
-            this.logger.Verbose("new comment added to article {ArticleId}, comment text: {CommentText}", articleId, comment.Text.Truncate(30, true));
+            this.logger.Verbose("new comment added to article {ArticleId}, comment text: {CommentText}", request.ArticleId, comment.Text.Truncate(30, true));
+
+            return new AddCommentResponse(true);
         }
 
         /// <summary>
         /// Удаляет комментарий.
         /// </summary>
-        /// <param name="articleId">Идентификатор статьи.</param>
-        /// <param name="commentId">Идентификатор комментария.</param>
-        public async Task DeleteCommentAsync(long articleId, long commentId)
+        /// <param name="request">Запрос удаления комментария.</param>        
+        public async Task<DeleteCommentResponse> DeleteCommentAsync(DeleteCommentRequest request)
         {
-            var article = await this.articleService.GetArticleByIdAsync(articleId);
-            var comment = await this.commentService.GetCommentById(article, commentId);
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            var article = await this.articleService.GetArticleByIdAsync(request.ArticleId);
+            var comment = await this.commentService.GetCommentById(article, request.CommentId);
 
             await this.commentService.DeleteCommentAsync(article, comment);
 
             await this.unitOfWork.SaveChangesAsync();
 
-            this.logger.Verbose("comment {CommentId} was deleted", commentId);
+            this.logger.Verbose("comment {CommentId} was deleted", request.CommentId);
+
+            return new DeleteCommentResponse(true);
         }
     }
 }
